requirements.txt
# external requirements
click
Sphinx
coverage
flake8
python-dotenv
pandas
numpy
tqdm
sklearn
nltk
transformers>=2.8.0
pytorch-lightning
wandb
tokenizers>=0.8.1
scikit-learn
scipy
allennlp
fairseq
fastBPE
nlpaug
textattack
pandas-bokeh
spacy
gensim
matplotlib
pyLDAvis
captum
eli5
sacremoses
subword_nmt
tensorflow_addons
tensorboard
sklearn_crfsuite
rouge-score